@@ -17,29 +17,25 @@
     "prepublishOnly": "npm run build"
   },
   "devDependencies": {
-    "@vitejs/plugin-react": "^1.3.2",
-    "eslint": "8.15.0",
+    "@vitejs/plugin-react": "^2.0.1",
+    "eslint": "8.22.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-react": "^7.30.0",
-    "microbundle": "^0.15.0",
+    "eslint-plugin-react": "^7.30.1",
+    "microbundle": "^0.15.1",
     "path": "^0.12.7",
-    "prettier": "^2.6.2",
-    "prettier-eslint": "^15.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "vite": "^2.9.9"
+    "prettier": "^2.7.1",
+    "prettier-eslint": "^15.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "vite": "^3.0.9"
   },
   "dependencies": {
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-env": "^7.18.10",
+    "@babel/preset-react": "^7.18.6",
     "@studio-freight/tempus": "^0.0.25",
     "nanoid": "^4.0.0",
-    "react-use": "^17.3.2",
+    "react-use": "^17.4.0",
     "throttle-debounce": "^5.0.0"
-  },
-  "peerDependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
   },
   "bugs": {
     "url": "https://github.com/studio-freight/hamo/issues"
