@@ -2,21 +2,22 @@ import React from 'react'
 import {
   useDebug,
   useDocumentReadyState,
+  useFrame,
   useId,
   useIsTouchDevice,
-  useRect,
   useIsVisible,
-  useFrame,
+  useRect,
 } from '../src/index.js'
 
 function App() {
-  const [ref, compute] = useRect(1000)
+  const [setRectRef, rect] = useRect(1000)
+  const { setRef, inView } = useIsVisible()
+
   const isTouch = useIsTouchDevice()
   const debug = useDebug()
   const ready = useDocumentReadyState()
   const id = useId()
-  const { setRef, inView } = useIsVisible()
-  const rect = compute()
+
   useFrame((time, deltaTime) => {
     console.log({ time, deltaTime })
   })
@@ -27,7 +28,17 @@ function App() {
       <p>is debug? {debug ? 'yes' : 'no'}</p>
       <p>is document ready? {ready ? 'yes' : 'no'}</p>
       <p>is in viewport? {inView ? 'yes' : 'no'}</p>
-      <div ref={ref}>top: {rect.top}</div>
+      <div ref={(node) => setRectRef(node)}>
+        top: {rect.top}
+        <br />
+        height: {rect.height}
+        <br />
+        left: {rect.left}
+        <br />
+        top: {rect.top}
+        <br />
+        width: {rect.width}
+      </div>
       <p>id: {id}</p>
     </main>
   )
